refactor(amazonq): move signalsAWSQDeveloperProfilesEnabled to qDeveloperProfiles
